[chore] 구조 변경
