--- app/src/main/java/org/sopt/dosopttemplate/presentation/LoginActivity.kt
@@ -1,39 +1,0 @@
-package org.sopt.dosopttemplate.presentation
-
-import android.app.Activity
-import android.content.Intent
-import android.os.Build
-import android.os.Bundle
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.appcompat.app.AppCompatActivity
-import org.sopt.dosopttemplate.data.UserInfo
-import org.sopt.dosopttemplate.databinding.ActivityLoginBinding
-
-class LoginActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityLoginBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityLoginBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        val resultLauncher =
-            registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-                if (result.resultCode != Activity.RESULT_OK) return@registerForActivityResult
-                val data = result.data ?: return@registerForActivityResult
-                val userInfo = getParcelable(data)
-            }
-
-        binding.btnLoginSignUp.setOnClickListener {
-            val intent = Intent(this, SignUpActivity::class.java)
-            resultLauncher.launch(intent)
-        }
-
-    }
-
-    private fun getParcelable(data: Intent) =
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            data.getParcelableExtra("user", UserInfo::class.java)
-        } else {
-            data.getParcelableExtra("user")
-        }
-}
